+++ packages/terra-clinical-detail-view/tests/nightwatch/DetailView/detail-view-spec.js
@@ -1,25 +1,16 @@
-/* eslint-disable no-unused-expressions */
-
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-clinical-detail-view', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a divided Detail View with provided components': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/detail-view-tests/detail-view-divided`)
+      .url(`${browser.launchUrl}/#/tests/detail-view-tests/detail-view-divided`)
       .assert.elementPresent('#DetailView hr');
   },
+
   'Displays a Detail View that is not divided with provided components': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/detail-view-tests/detail-view-no-divider`)
+      .url(`${browser.launchUrl}/#/tests/detail-view-tests/detail-view-no-divider`)
       .assert.elementNotPresent('#DetailView hr');
   },
-};
+});